backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
django-cors-headers==4.6.0
python-dotenv==1.0.0
pyotp==2.9.0
bcrypt==4.1.2
cryptography==41.0.7
Pillow==10.1.0
django-csp==3.7
whitenoise==6.6.0
dj-database-url==2.1.0 